wondershaper: Fix definition of local traffic
Prior to this commit local traffic was defined as any traffic that
explicitly mentioned the traffic shaping device's local IP (e.g.
192.168.1.1). However, this is a bad definition as internet traffic that
goes through an ADSL routing that applies NAT will make the packets
explicitly state they are for this local IP.

This patch narrows the definition to packets which mention the traffic
shaping device's local IP *and* the other end is a local subnet (e.g.
192.168.x.x).