Enhance INTERNET->LAN rates
To achieve a target latency of 50ms (1/20th of a second) it is important
to take preventative action before the target is reached. This patch
changes the burst rates to half the target (25ms, or 1/40th a second)
and sets the cburst to the size of a typical packet (or "average packet"
as the `man tc-htb` calls it).

Experiment has shown that these changes keep ping latency to ~80ms when
the DOWNLINK is completely congested. (Ping is ~20ms when DOWNLINK is
otherwise idle). While this is not quite the target of 50ms, it is "good
enough".